+++ include/Globals.h
@@ -1,6 +1,8 @@
 #pragma once
 
 #include "FastMutex.h"
+#include "ExecutiveResource.h"
+#include "LookasideList.h"
 
 /**
  * @brief Holds all the global state of the driver, instead of creating separatge global variables
@@ -10,7 +12,7 @@ struct Globals
     /**
      * 
      */
-    VOID Init(_In_ ULONG maxCount);
+    VOID Init(_In_ ULONG maxCount, _In_ ULONG processesMaxCount);
 
     VOID Destroy();
 
@@ -29,9 +31,31 @@ struct Globals
      */
     LIST_ENTRY* RemoveItem();
 
+    /**
+     * 
+     */
+    bool AddNewProcess(_In_ LIST_ENTRY* entry);
+
+    /**
+     * 
+     */
+    bool RemoveProcess(_In_ HANDLE pid);
+
+    /**
+     * 
+     */
+    VOID ClearNewProcesses();
+
 private:
+    // Items 
+    // Process, Thread, Image & RemoteThread
     LIST_ENTRY _ItemsHead;
-    ULONG _Count;
-    ULONG _MaxCount;
-    FastMutex _Lock;
+    ULONG      _Count;
+    ULONG      _MaxCount;
+    FastMutex  _Lock;
+    // New Processes
+    LIST_ENTRY        _NewProcessesHead;
+    ULONG             _NewProcessesCount;
+    ULONG             _NewProcessesMaxCount;
+    ExecutiveResource _NewProcessesLock;
 };